+++ lang/syn/src/parser/program/instructions.rs
@@ -4,6 +4,7 @@ use crate::parser::spl_interface;
 use crate::{FallbackFn, Ix, IxArg, IxReturn, Overrides};
 use syn::parse::{Error as ParseError, Result as ParseResult};
 use syn::spanned::Spanned;
+use syn::Attribute;
 
 // Parse all non-state ix handlers from the program mod definition.
 pub fn parse(program_mod: &syn::ItemMod) -> ParseResult<(Vec<Ix>, Option<FallbackFn>)> {
@@ -28,12 +29,14 @@ pub fn parse(program_mod: &syn::ItemMod) -> ParseResult<(Vec<Ix>, Option<Fallbac
             let overrides = parse_overrides(&method.attrs)?;
             let interface_discriminator = spl_interface::parse(&method.attrs);
             let docs = docs::parse(&method.attrs);
+            let cfgs = parse_cfg(method);
             let returns = parse_return(method)?;
             let anchor_ident = ctx_accounts_ident(&ctx.raw_arg)?;
             Ok(Ix {
                 raw_method: method.clone(),
                 ident: method.sig.ident.clone(),
                 docs,
+                cfgs,
                 args,
                 anchor_ident,
                 returns,
@@ -145,4 +148,15 @@ pub fn parse_return(method: &syn::ItemFn) -> ParseResult<IxReturn> {
             "expected a return type",
         )),
     }
+}
+
+fn parse_cfg(method: &syn::ItemFn) -> Vec<Attribute> {
+    method
+        .attrs
+        .iter()
+        .filter_map(|attr| match attr.path.is_ident("cfg") {
+            true => Some(attr.to_owned()),
+            false => None,
+        })
+        .collect()
 }